+++ lib/show_for/builder.rb
@@ -67,7 +67,7 @@ module ShowFor
       if tag
         type_class = ShowFor.send :"#{type}_class"
         html_options = options.delete(:"#{type}_html") || {}
-        html_options[:class] = "#{type_class} #{html_options[:class]}".rstrip
+        html_options[:class] = [type_class, html_options[:class]].compact.presence
         @template.content_tag(tag, content, html_options)
       else
         content